Refactor socket
